Extract cancelSync method call to utils to stub the call more easily
